Adding Music Project
I am adding the project that Patrick started for me. This project at this point is most his code and contains his audio files. I am soon going to change the looks, add functionality and replace the audio files with my own, making the app actually fit for its intended purpose. There are a few things here that are my code such as the link to the chord chart, which is a chord chart I made, and there are a few other minor changes that I made.